% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_pkgs.R
\name{link_pkgs}
\alias{link_pkgs}
\title{Convert a List Column of Packages to Character Columns with Links}
\usage{
link_pkgs(x, remove = character())
}
\arguments{
\item{x}{(\code{list()})\cr
List column.}
}
\description{
Currently only supports linking to CRAN packages.
}
