% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3website-package}
\alias{mlr3website}
\alias{mlr3website-package}
\title{mlr3website: R Package to Build the Website of the mlr3 Ecosystem via Quarto}
\description{
Use this package to install helper files required to build the mlr3 website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr-org/mlr3website}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3website/issues}
}

}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID})

}
