% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{mlr_pkg}
\alias{ref}
\alias{cran_pkg}
\alias{gh_pkg}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{mlr3book}{\code{\link[mlr3book]{cran_pkg}}, \code{\link[mlr3book]{gh_pkg}}, \code{\link[mlr3book]{mlr_pkg}}, \code{\link[mlr3book]{ref}}}
}}

