% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_responsive.R
\name{table_responsive}
\alias{table_responsive}
\title{Convert Tabular Object to HTML Table}
\usage{
table_responsive(tab, ...)
}
\arguments{
\item{tab}{(\code{any})\cr
Tabular R object.}

\item{...}{Passed to \code{knitr::kable()}.}
}
\description{
Converts a tabular R object with \code{knitr::kable()} to a html table. Wide
tables can be scrolled horizontally. The table is scrolled vertically if its
length exceeds 700px. The table header stays fixed at the top when scrolling
down. The following CSS rules must be applied.

\if{html}{\out{<div class="sourceCode css">}}\preformatted{/* Scroll container around table*/
.table-responsive \{
  width: 100\%;
  max-height: 700px;
  overflow: auto;
\}

/* Table is stretched to full width */
.table \{
  width: 100\%;
\}

/* Sticky table header and background color */
.table > thead \{
  color: white;
  background-color: #495148;
  inset-block-start: 0;
  position: sticky;
\}

/* Display line breaks */
.table td \{
  white-space: pre-line;
\}
}\if{html}{\out{</div>}}
}
